move codec functions to codec.ts; change test file name; export all in index.ts
